Create header, unit links, target text and sidebar
Create the layout for QuickQWERTY home page. The header consists
of three columns:

  1. QuickQWERTY logo
  2. Unit
  3. Page links

Below the header, there is a three column layout for the
following items:

  1. Unit links
  2. Practice panel
  3. Sidebar

Make the unit links static for the time being. Similarly, add a
static target text in the practice panel. Let the sidebar
contain a sample text.