Remove thin gray box that appears if there are no subunits
When there are no subunits within the div element with
subunitLinks class, a small gray box appears for the div element
with .stretch class due to two issues:

  1. The 'width: 19%' property in '#subunitLinks div' overrides
     'width: 100%' property in '.stretch'.
  2. The 'border' property in '#subunitLinks div' creates a thin
     gray border.

We want the width of the div element with .stretch class to be
100% while the div element itself should be completely
invisible. Fix this by increasing the specificity of the
selector of this element from '.stretch' to
'#unitLinks div.stretch'. Remove the gray box by explicitly
defining 'border: 0' property for the selector of this element.