Dynamically generate subunit links from units.js
The subunit links are currently static div elements in HTML.
Remove these static div elements from HTML. Generate these links
dynamically for Unit 1 only.

Ensure that the function that generates the links dynamically
accepts an integer argument with which the unit number for which
the subunit links are to be generated can be specified. Invoke
this function for Unit 1 at this time.