Dynamically generate target text from units.js
The target text consits of three parts:
  1. Prefix
  2. Target character
  3. Suffix

The target character is the character the user should type to move
ahead in the subunit. The prefix and the suffix offer some context
around the target character to be typed. These three parts combined,
in the specified order above, is a substring from the subunit's text
from units.js.