Create a global object to contain tutor properties
The code in tutor.js contains a few global variables. This makes
finding the declaration of each global variable in use
difficult. Create a global object called 'tutor' to solve this
problem. Turn all the existing global variables into properties
of this global object. Thus, wherever a property of 'tutor'
object is used, the property would be prefixed with 'tutor.'
thereby making it clear that the property belongs to the global
'tutor' object.