Disable spell check in input pane
The textarea HTML element in the input pane has spell check
enabled by default. As a result, when the user types the target
text, the browser displays red underlines under the typed text.
This could confuse the user into believing that he or she has
made an error in typing. Therefore, disable spell check in this
textarea element.