Set background of input textarea to pink on error
The background of the target text element is set to pink when
the user makes an error. However, the background of the input
textarea field remains white. Change the background of the input
textarea to red as well when an error occurs.

Further, do not make the text of the input textarea element red
on error since red text on pink background is hard to read. Let
it remain gray.