Update progress, speed and error as the user types
Make QuickQWERTY a working application by making the following
elements functional:
  1. Progress bar
  2. Progress percent
  3. Typing speed in CPM
  4. Typing speed in WPM
  5. Error rate in percent

To make the above elements functional, update them after every
keystroke by the user. To compute WPM, a word is standardized to
five characters.

Since the application is supposed to become functional with this
commit, remove any text from the status pane and sidebar which
are not functional yet.