Add links to go to previous and next subunits
Provide two links on either sides of the practice pane to go to
the previous subunit and the next subunit. If the user is
already at the first subunit, clicking the previous link should
do nothing. If the user is already at the last subunit, clicking
the next link should do nothing.