Reduce the height of each unit link
Currently, the height of each unit link is 24px. It was found
that in a screen of resolution 1280 x 800, all the unit links do
not fit in a screen, and as a result one has to scroll down to
click Unit 21. Therefore, reduce the height of each unit link to
21px, so that the entire list of unit links fit in a screen and
the user doesn't have to scroll down to click the last unit
link.