Add a typing help link to show or hide typing tips
Create a link called 'Help' that the user can click to see tips
on how to place the fingers on the keyboard and type in the
current unit. Once the help is shown, the link changes itself to
'Hide Help' which can be clicked to hide the tips. By default,
when the page loads, the tips are hidden.