Display error message if JavaScript is disabled
If JavaScript enabled browser is not found, display a message
that JavaScript should be enabled. Also, hide the practice pane
by default so that it is invisible when a JavaScript enabled
browser is not found. Make the practice pane visible from
JavaScript code, so that it is displayed to the user only if the
user is running this application with a JavaScript enabled
browser.