Let 4 pictures fit per row in the Facebook widget
The Facebook like box on the home page displays only 3 profile
pictures per row. The width of the widget was set to 220px
earlier. Increase the width to 239px, so that it can now contain
4 profile pictures per row.

The width value of 239px was arrived at by doing an analysis of
why the default Facebook widget that is 292px wide can contain 5
profile pictures per row. Each profile picture is 50px in width
and has a right margin of 3px. There is a padding of 10px on
both sides. There is a border 1px wide on both sides.

    (50 + 3) * 5 + (10 * 2) + (1 * 2) = 287

Therefore, the Facebook widget should be at least 287px wide in
order to contain 5 profile pictures per row. It seems like
Facebook has added 5px to this value to arrive at the default
width of 292px.

Similar calculation shows that the widget must be at least 234px
wide in order to contain 4 profile pictures per row.

    (50 + 3) * 4 + 10 * 2 + 1 * 2 = 234

After adding an extra 5px, the width turns out to be 239px.