Remove semicolons from anonymous function calls
In this project, by convention, semicolons are not used at the
end of JavaScript statements. However, the anonymous function
calls that define the Tutor and Util objects are terminated with
semicolons. Remove these semicolons for the sake of consistency
in order to conform to the coding convention used in this
project.