Add 5-6 split units for number keys (Units 16-20)
In most keyboards, the '6' key is located closer to the 'F' key.
As a result, the '6' key is more accessible to the left hand
than the right hand. Therefore, many people prefer the
traditional style with 6-7 split where '6' is pressed with the
left hand and '7' is pressed with the right hand. However, there
are also many people who prefer the nontraditional style with
5-6 split. Therefore, add support for 5-6 split typing for users
who prefer this by providing the following features:

  1. Add a link below the input box called '6-7 split' or
     '5-6 split' only for units which introduce the number keys
     to the user, i.e. units 16-20.
  2. When a user clicks on one of these links, the corresponding
     unit is loaded.
  3. Update README.html with information about the alternate
     units.
  4. Fix README.md so that it mentions units 16-20 as the units
     that introduce the number keys.