Add footer, tutor link and descriptive page titles
Add a footer to each page. The footer contains navigation links
along with a link to the source code on GitHub and a copyright
notice. Add a link to the home page, i.e. the tutor in the
header as well as the footer. Edit the title of each page, so
that each title is distinct and descriptive.