Make the padding around main element uniform
Currently the bottom padding around main element is 3% while the
top, right and left padding is 2%. This causes the page layout
to look asymmetric when the header as well as the footer fits in
the same screen, i.e. both header and footer can be seen without
having to scroll. Make the padding 2% for all sides so that the
layout looks symmetric.