On URL parsing error, display unit 1 or subunit 1
If a user visits, http://quickqwerty.com/#restart,
http://quickqwerty.com/#next, etc., then the practice panel does
not appear due to a JavaScript error. This issue occurs because
the fragment identifier, if present, in a URL typed by the user
is expected to be either in M or M.N format where M is the unit
number and N is the subunit number. To fix the issue, when the
integer M, or the dot separated integers M and N are not found,
assume M = 1 or N = 1 for the missing values.