Centre noscript message
The noscript message that appears when JavaScript is disabled is
not vertically aligned at the centre. This is due to the active
practice panel elements that are empty but occupy space above
the noscript message. Disable all practice elements by default,
and modify the HTML layout so that the noscript message appears
to be centred.