Make Facebook like box visible during offline use
The Facebook like box is not visible during offline use. This
happens due to a Facebook SDK bug due to which it is unable to
set the width and height properties of the like box elements.
Therefore, set these properties using CSS to work around this
bug and display the like box.