Fix bug to display version number in footer
In the 'License' and 'About' pages, a JavaScript bug prevents
the version number of QuickQWERTY from being displayed in the
footer. The bug is due to Tutor related code in these two pages.
It's a bug because the Tutor object is undefined in these two
pages. Remove this Tutor related code from both pages to fix the
bug.