Remove additional space below help text
There is a huge amount of vertical space between the help text
and the footer. This is due to the margin around noscriptWrapper
element that is outside the noscript tags. As a result, the
noscriptWrapper element is active and the margin around it
pushes the footer further down even when JavaScript is enabled.
Fix this bug by moving the noscriptWrapper element into the
noscript element.