Add input commands: restart, reset, fix and xxx
Process command entered by the user in the input textarea
element when the tutor is in error state. Once the tutor enters
the error state, the supported commands are searched at the end
of the erroneous input. If a supported command is found at the
end of the erroneous input, the command is processed.

The following is a list of the supported commands:
  restart  restart the current subunit
  reset    same as the 'restart' command
  fix      remove errors from the input text
  xxx      same as the 'fix' command