Add version update check feature
Check if there is a version update required. If a stable version
is available that is newer than the version of the application
the user is using, then display a message with the details about
the version update and allow the user to update or ignore the
update.

The update message is displayed to the user only when a new
session begins until the user hides the message. If the user
hides the message, then the update message is not visible
anymore till the end of the session, and it is visible again
from the beginning of the next new session.

Mark the current stable version as 0.4.1.