Add a download page
Create a page called download.html. This page displays the
following information:
  1. Link to download the latest stable version
  2. Link to download the latest development version
  3. Link to use QuickQWERTY online
  4. Link to all releases on GitHub.

Since no stable release has been made yet, the first section to
display the link to the latest stable version should mention
that no stable release is available. When this page is not
accessed from quickqwerty.com, a warning should be displayed
that this page may be outdated.

Update the download URL in the headers of all pages to point to
this new page.