Fix: Focus is not set on input box on page load
When the tutor is loaded or reloaded, the focus is not set on
the input textarea element. This is due to a bug in the code
that initializes the tutor. The focus is set on the input
textarea element before the practice panel wrapper containing
the textarea element is displayed. Fix this bug by setting the
input textarea element after the practice panel wrapper is
displayed.