Do not update tooltips on every key press
When a user is typing the current lesson, the user is not likely
to pause and have a look at the progress, speed and error
tooltips. Therefore, do not update these tooltips when the user
is typing. Update the progress tooltip only once when a new
lesson is loaded. This tooltip should display the number of
characters in the lesson. Update speed and error tooltips only
after the current lesson is completed.

Set Version.VERSION = '1.0-beta2.3'.