Remove redundant code to select unit and subunit
There is a function called selectUnitAndSubunit in Tutor object
that is used to select the current unit and subunit. However,
the current unit and subunit are already marked as selected
using the boxedLink function. Therefore, the
selectUnitAndSubunit function is no longer required. Remove this
function as well as call to this function.