Previous/next link should redirect to subunit URL
When the user is at the first lesson and presses the previous
link, or when the user is at the last lesson and presses the
next link, the URL changes to /#previous or /#next and remains
so. Change the behaviour such that the URL /#previous redirects
the user back to the URL of the first lesson when the user
clicks the previous link while he is at the first lesson.
Similarly, /#next should redirect the user back to the URL of
the last lesson when the user clicks the next link while he is
at the last lesson.

Set Version.VERSION = '1.0-beta2.4'