Remove large download button from 'Download' page
Remove the level 2 heading displayed as a download button from
the 'Download' page, i.e. download.html. This button is a
redundant button that is a larger form of a download button
already present in one of the release sections in this page.
Instead, replace this button with a normal level 2 heading and a
short introduction to offline usage of QuickQWERTY.