Run completion tasks after updating result pane
The speed submitted via OnePixel log and the speed reported in
the tooltip is the speed calculated after typing the last but
one character of a lesson. The speed calculated after typing the
last character is displayed only in the result pane. This is
neither submitted via OnePixel log nor displayed in the tooltip.
This issue occurs because the log is submitted and the tooltip
is updated before the speed and error are computed after each
key press. Fix this issue by submitting the log and updating the
tooltip after the speed and error are computed after each key
press.