Do not display 'Next lesson' link for last subunit
When a user completes the last subunit, i.e. Unit 21.3,
satisfactorily, the tutor displays 'Next lesson' as the advice.
This is a bug because there is no next lesson when the user is
at the last subunit. Therefore, fix this bug so that the 'Next
lesson' advice is not displayed when the user completes the last
subunit satisfactorily.