Set focus on advice link on completing a subunit
When a subunit is completed, an advice link appears in the
status pane to let the user repeat the current subunit or go to
the next subunit. Set the focus on this link after it appears so
that the user can click this link simply by pressing the 'Enter'
key.

Set Version.VERSION = '1.0-beta2.5'