Fix typos and format license text in code
Every HTML and JS file have the license text as comments in the
code. Perform the following changes in the license text.

  1. Fix typos.
  2. Insert a blank line before the list of conditions in order
     to neatly separate it from the previous paragraph.
  3. Indent the list by two spaces only instead of four.

These changes make the license text conform to Markdown syntax.
If the license text is now converted to HTML using a Markdown
converter, the list of conditions would appear as an ordered
list that is neatly separated from the previous paragraph.