Use Tutor.VERSION instead of Version.VERSION
In the previous commit, version.js that defines the Version
object was removed. The version information was moved into Tutor
object in tutor.js. Therefore, edit all HTML and JavaScript
files to use this new version in Tutor.VERSION.

Also, do not update the version displayed in the footer
dynamically using JavaScript. Instead, mention the current
version as static text in the footer.