Add script to perform automatic substitutions
QuickQWERTY is a project that contains static webpages. The
static webpages contain several pieces of text that need to be
updated from one release to another. For example, copyright
notice in each file need to be updated with the current year of
publication every year, version mentioned anywhere in the pages
need to be updated on every release, certain sections in a file
may need to be included or excluded depending on the release
cycle, etc.

A script dev/autosub has been written to automate such updates,
inclusions and exclusions.